+++ src/api/balance/model.ts
@@ -1,0 +1,4 @@
+export type BalanceModel = {
+  balance: number;
+  updated_at: string;
+};
