+++ src/api/upload/model.ts
@@ -1,0 +1,6 @@
+export type UploadModel = {
+  url_id: string;
+};
+export type UploadFileModel = {
+  file: File;
+};
